Move system to lib
